Updated the dbService code with fetchAndSaveStaticData() exposed
